Transform function expression in expression macro instead of build macro
At that point, all fields have been typed and that information can be
used, as with `Context.typeof` in `ETry` expression handling.

This fixes the `ETry` handling, that was partially broken.  It used
`Context.typeof` before the field was typed and, therefore, didn't
recognize valid identifiers such as local variables or other fields.
The following example illustrates what failed with
    `Unknown identifier : bar`
before and now works:

    @:build(com.dongxiguo.continuation.Continuation.cpsByMeta("async"))
    class Test {
        @async static function foo(bar:String):Int
        {
            var x = try {
                // sync stuff only, obviously
                bar;
            } catch (e:Dynamic) {
                return 1;
            }
            // some async stuff
            return 0;
        }
        static function main()
        {
            foo("hi", function (code) trace(code));
        }
    }